+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.4.0 - 2017-10-01
 
+- Support mixed geojson `closest-end-nodes`
 - Update TurfJS dependencies
 
 ## 0.3.0 - 2017-09-29